fix some issue with grouping and add to changlog
