+++ packages/cli/src/commands/split.ts
@@ -2,8 +2,9 @@ import path from "path";
 import fs from "fs";
 import { getDependencyTree } from "../helper/dependencies";
 import { cleanupOutputDir, createOutputDir } from "../helper/file";
-import { getGroupsFromTree, splitPath } from "../helper/tree";
+import { getEndpontsFromTree, splitPath } from "../helper/tree";
 import { GroupMap } from "../helper/types";
+import { getGroupsFromEndpoints } from "../helper/groups";
 
 export default function splitCommandHandler(
   entrypointPath: string, // Path to the entrypoint file
@@ -17,8 +18,11 @@ export default function splitCommandHandler(
   cleanupOutputDir(outputDir);
   createOutputDir(outputDir);
 
-  // Iterate over the tree and get groups
-  const groups = getGroupsFromTree(tree);
+  // Iterate over the tree and get endpoints
+  const endpoints = getEndpontsFromTree(tree);
+
+  // Get groups from the endpoints
+  const groups = getGroupsFromEndpoints(endpoints);
 
   // Process each endpoint for splitting
   for (const group of groups) {
